First draft of new DSL interfaces.
This includes the user model (flavor/build types),
part of the variants (lots more needed), and the
extensions.

This also starts doing some implementations of the
external objects with the sealing mechanism.

In order to implement these this commit provides
sealable implementations for List/Set/Map and
all the other intermediary classes that they
need (Iterator, ListIterator, as well as sets
and collections for Map.entries/keys/values)

Test: N/A
Change-Id: Ia710275f2bfd877acb0fdc92681805f3e6828c15

Former-commit-id: d95fb1d71d8b82b9b058942dfdb4cfcc7b1365a6