Register transform streams as separate inputs
We now register each consumed and referenced stream
as a separate input with relative path sensitivity.
Because each stream is a file tree, every change will
be relevant only in the context of the stream that owns
the parent directory.

Previously, we would register all streams using a single
getter annotated with @InputFiles with relative path
sensitivity i.e. we would register a single task input.
This means that all incremental changes would be evaluated
relative to all input roots. So, if a file com/test/X was
removed from input I, but a new file com/test/X was added
to input J, incremental input data would report this as
changed file in stream J.

BUG=65363841
Test: DexArchivesKotlinTest

Change-Id: I17963891ad7b72546cf64c1f2d2fba0ae817b00d

Former-commit-id: ae35d1957b44006b727f9aa18613c5e087dad2b7