Write failed AAPT2 JNI command to file
AAPT2 JNI commands can be very long since we pass all the resources by
path one by one. If a command failed we'd write it as part of the
exception resulting in a lot of spam and not much good information to
the user. This CL saves those arguments to a file so we still have
access to it, but saves a lot of clutter for the user.

Test: Aapt2V2JniTest
Change-Id: I86f7419c254e0f500cd0ff6218a773446ad7e42d

Former-commit-id: a360c4347f98836e08b86e58e35e91f9c8ed2fc8