Adding handling for L and M InputManger events.
With L and M the interface for the InputManager was one way,
in N it was changed to be another way. This change looks for
fields that are available in N, O however if they are not found it
continues to look for fields available in all versions. When it finds
that field, it will check to see what type of object we have and call
the proper functions to get access to the base connection for us to
wrap.

Change-Id: I490464ae43ae17e48c662042bfb905379e8ba180
Fixes: 64959834
Test: BasicPerfTest, will test the N+ path, the pre-N path will come later

Former-commit-id: 01a4cee450c89e1cd72f424ab931645224b8b86d