Change name of directory that contains merging log files.
Previously, these log files used to be kept under the
"intermediates/blame/../multi" directory, and we did not clean it at the
start of a full build.

Recently, we introduced a new format for these log files (commit
d2095ddfc618604738bc5e9d6a60928a7b4421d0 [formerly 4ef3d2499148d6e32d6727ed78cd411da197975b]). Therefore, if the users
switch between versions of the plugin that have different formats of the
log files, the next build will fail to read the existing log files. (We
tried to support a newer plugin version reading the old format, but we
still cannot support an older plugin version reading the new format.)

To work around this issue, we need to use a different directory called
"intermediates/blame/../multi-v2" to contain log files with the new
format.

Note that, with a recent change (commit
6c3a54f51e58d2f1ce26ec8255fa374da9883afc [formerly da519f86eb84477ef2823dfec27b984392fa3fa5]), we now also clean these log
files at the start of a full build. Therefore, moving forward, we won't
need to use a new directory again even if we introduce yet another new
format.

Also, as it is no longer necessary to support reading the old format,
this commit removes that part of the code as well.

Bug: 65263212
Test: Updated MergingLogTest
Change-Id: Ia5837b940d10a7a8b3686029b4986cce98093771

Former-commit-id: e9492eb028841e1041dfd3d213c4aa0e8af3b18e