Use project instance as key for plugin version check.
The recently introduced plugin version consistency check requires that
different projects (root projects, subprojects, or included projects)
must apply the same version of the plugin within a build.

It additionally requires that a project must apply the plugin only once
within a build.

Previously, we identified a project with the project path. However,
within a build, Gradle might apply the plugin multiple times to
different project instances having the same project path. In that case,
our check would fail where it should not.

To fix this, we now use the project instance instead of the project path
to identify a project. We still maintain the two checks above.

Bug: 65502981
Test: Updated PluginInitializerTest
Change-Id: If73fcd46fac2af1ea46f99cafa80ddd4bae1ff1a

Former-commit-id: 0df2ad4d8260aa14d16086b08479b22e93779734