Enforce loading the plugin only once within a build.
If the buildscript classpaths that contain the Android Gradle plugin in
sub-projects, or included projects in the case of composite builds, are
set differently, the plugin will be loaded in different classloaders.

In that case, it will break Gradle's variant-aware dependency management
as Gradle fails to compare attributes of the same class but loaded by
different class loaders.

For now, we need to detect this issue and fail the build early.

The suggested fix is to add the Android Gradle plugin to only the
buildscript classpath of the top-level build.gradle file. In the case of
composite builds, also make sure the buildscript classpaths that contain
the Android Gradle plugin are identifical across the main and included
projects.

We added a flag (android.enableBuildScriptClasspathCheck) to let users
disable this check in case they are using a version of Gradle that has
fixed this issue.

This commit also removes the composite build warning in BasePlugin,
since it is subsumed by this check.

Bug: 64978740, 65393481
Test: Updated PluginInitializerTest, BuildSessionImplTest

Change-Id: I685e586d3758f0ac5fcfa56eb0ab409a7ebaca86
(cherry picked from commit ebde37d174719ae6cc4cf6abebb01b5c1215207e [formerly 66d50cec99f6b7e3d89b7a4e77c3565ff3277227])

Former-commit-id: 29008d33819806cab0d72fd85d97aad5767d0a2e