Use D8 to get the main dex list
Use D8 to calculate the main dex list. The
tool handles the runtime annotations, so we do
not need to handle that ourselves. Inputs given
to the tool are the program classes, library classes
(e.g. android.jar), and the hardcoded rules. Output is
a list of classes to keep.

BUG=65043103
Test: D8MainDexListTransformTest, MultiDexTest,
DesugarAppTest, NewMultiDexMainDexListTest

Change-Id: I28832184592ef401df0d1c92b24da590f0074c9a

Former-commit-id: 18036dc458dc26523b081630782ebe0c33f30571