Show better CMake error messages
Capture all the CMake error messages (if and when they occur) and show
that to the user instead of showing the very last CMake error message
which might not necessarily be helpful.

Test: Use a test app and intentionally introduce errors (invalid params)
etc and see if the output shows them.

cmake-error-msg

Change-Id: I734fb236b48a3adca060083b14a034f7b895543b
(cherry picked from commit Ie197c8048287a753db34457ec383cf9c432ab408)

Former-commit-id: 47bf5dc1623f44229757ec5e474ed3410bc98e82