65549795: NewApi violations not detected on android.view.View
This removes ApiLookup#isRelevantOwner and replaces it with an
instance methods that actually consults the API database.

However, some of the custom lint check unit tests (loading
custom jar files containing lint checks.) These had to be
updated, and I took the opportunity to rename the id for one
of those custom checks since it had the same id as one of
the checks that are now built-in and this has been confusing
in the past.

Inspect: OFF
Bug: 65549795
Test: Unit test included
Change-Id: I9926c7e9e01da875ed6aacabd61da20cf3510d0d

Former-commit-id: b97f99c3f68b8caa82ff1aa84aa9e8430f27355f