Update file_cache constructor to take path instead of boolean
Previously, the file_cache constructor changed due to a temporary new
approach to tests that has since been changed further. This allows us to
simplify the API, reducing the number of constructors back down to 2
(a normal constructor and a dependency injection constructor) and
removing any concept of special-casing logic for tests.

Test: N/A, refactoring only. Covered by existing tests.
Change-Id: Ib05e46aeece310face7cfba8a5268512b4bff0ab

Former-commit-id: 38031c2deef0d52b0fb20503c10ff064b5978daa