Fix clip path within a style attribute
When clip path is defined as part of the style, we will handle it as if
it is an attribute of clip-path.
We need to keep track of the path name early when we build the map of
clip path, for its name to its parent node. Since we don't want to
traverse the whole tree one more time.

b/67408766
Test: added new tests

Change-Id: Ic8dc0bbad656dd6ccf1f8da6c9bd68038949fb27

Former-commit-id: fccaa0175222241b7c71a2456a9daef55542db65