Reduce CMake tests on Windows because they are too slow
We plan to improve CMake sync performance on Windows
but it will take some time. In the meantime, disable some
time-consuming tests on Windows.

For STL matrix test, reduce to just libc++ (static and dynamic)
because those will be the only STLs left shortly. This should
divide time of test by three.

Test: this is a test-only change

factor-native-model-test

Change-Id: Ife7dd7a870dc722e77b699e908c8e3c79a739682

Former-commit-id: b40a39f2d45719d0fcc3f5a07e2318c8672daea9