Generate base binding classes in gradle
This CL adds a new gradle.property called android.databinding.enableV2
which enables a new compiler setup in data binding.

In that setup, generated binding classes that are accessed by user code
are created with the Data binding source gen task rather than the
annotation processor. This works around a problem in javac where if
the compilation fails for any reason, it would print an error for
each place user accesses this generated code, practically hiding the
actual error in a moderate size project.

Bug: 68393034
Test: DataBindingExternalArtifactDependencyTest, DataBinding integration tests
with V2
Change-Id: I1307664af30a142b308b98388549d776a269f655

Former-commit-id: a062f7e367c90203e95086c0b46c1e8bd25a708b