Fix perf tests.
I made a change that switched the ordering of project configuration and
getting the executor from a project, such that that AntennaPod tests
were getting an executor for the root project instead of a subproject.
This was causing Gradle to not be able to find "assembleDebug" and
other such tasks.

Change-Id: I9d05013d2b525a723bc54c0c1c58b2b871253e6c
Test: ./gradlew performanceTest

Former-commit-id: e804d86e81e1d6d7a85df654f4afd01366a9e1fc