Refactor and improve Desugar workers
There are two related changes in this CL:

- reuse the logic for evaluating the desugar
process arguments. This is now extracted in
DesugarProcessArgs class, and shared between
worker and non-worker implementation.

- For Desugar Gradle workers, use the same
logic for bucketing inputs and outputs as for
the non-worker use-case. Lambda tmp dir is
changed for every Desugar invocation, so
worker processes was created for every input jar.
Now, only number of buckets Desugar processes is
created, which is number of cores.

BUG: 69589364
Test: existing Desugar tests
Change-Id: I7c3f58f2925d81476bb1c8ee4d5c1eea4850d89b

Former-commit-id: a6f4a24d2b85cbf51c2a78fba50eb951db5abd78