More granular Desugar incremental processing
When only a few class files change, copy them
to a temporary directory, and trigger desugarring.
This avoids processing all classes.

Because there is I/O cost to copying the files,
this will be done only if directory input is
significantly large, and if less than 10% class
files chaged.

In order for incremental builds to be correct,
all class files in the same package that start
as the changed file will be reprocessed, and all
class files in the output from the same package
that start with the same name will be delted.

BUG: 70326361
Test: DesugarTransformTest
Change-Id: Ic62b834b28ad4e85a4c94e07ae343df4d203da44

Former-commit-id: c82a4872b580eee42ea6c71df081527c845f982b