Refactor LocalUploader.
I want to add a new profile uploader that dumps the raw protos locally,
so I'm renaming this to clarfiy its purpose and so that it doesn't
clash with the class I'm planning to add.

While I'm here, I also made it more in line with ActdProfileUploader
and removed the need to mock it for testing.

Test: existing
Change-Id: Iddc6ff4acfa2bf5cb202a7b653f6ff052783a6bd

Former-commit-id: 070b02aaddff79001e8c7a740b121127c2c64662