Create a local proto profile uploader.
The purpose here is to save the protos generated by the benchmarking
runs to local disk if the user so chooses.

My plan is to use this to generate a representative set of benchmark
result protos for a single run to then look in to the uploading of those
to Dana.

Test: LocalProtoProfileUploaderTest
Change-Id: Ibc34d5ff96511f544643c84b9dff11876934136a

Former-commit-id: 242d555c449fd03e3b78def50a740e8695e6f0f0