Make defaultUploaders() more friendly.
In its previous incarnation, this method would hard-fail when one of the
uploaders couldn't be created. Commonly, the ActdProfileUploader would
fail because one of its environment variables wasn't set correctly.

This is a trade-off: making this method more permissive for the local
use-case makes it such that errors on the builders could be missed. I'm
open to better ideas that make it easy to run locally and hard to screw
up in the builders.

Test: existing
Change-Id: I41c6ee7a6006c3f64dda322e85b7536d1a5fc75b

Former-commit-id: ff64f1b79a0aec951d9174a4cbd27bc619842302